if fitting aborted, cluster can continue running

+++ estimate_explosion_time/cluster/__init__.py
@@ -3,15 +3,19 @@ import time
 import subprocess
 import logging
 import numpy as np
+from tqdm import tqdm
 from estimate_explosion_time.cluster.make_cluster_submit_script import make_desy_submit_file
-from estimate_explosion_time.shared import log_dir, cache_dir, get_custom_logger, main_logger_name
+from estimate_explosion_time.shared import log_dir, cache_dir, get_custom_logger, main_logger_name, TqdmToLogger
 
 
 logger = get_custom_logger(__name__)
 logger.setLevel(logging.getLogger(main_logger_name).getEffectiveLevel())
+tqdm_deb = TqdmToLogger(logger, level=logging.DEBUG)
+tqdm_info = TqdmToLogger(logger, level=logging.INFO)
 
 
 username = os.path.basename(os.environ['HOME'])
+cmd = f'qstat -u {username}'
 
 
 def submit_to_desy(method_name, indir, outdir, njobs, cache=cache_dir, simulation_name=None):
@@ -20,9 +24,8 @@ def submit_to_desy(method_name, indir, outdir, njobs, cache=cache_dir, simulatio
 
     # remove old logs
     logger.debug('removing old log files')
-    for file in os.listdir(log_dir):
+    for file in tqdm(os.listdir(log_dir), desc='removing old log files', file=tqdm_deb, mininterval=30):
         if name in file:
-            logger.debug(f'removing {log_dir}/{file}')
             os.remove(log_dir + '/' + file)
 
     submit_file = make_desy_submit_file(method_name, indir, outdir, cache)
@@ -57,27 +60,20 @@ def wait_for_cluster(job_id):
     on cluster, and outputs full table result every ~ 8 minutes. On
     completion of job, terminates function process and allows the script to
     continue.
+    :param job_id: int, the ID of the job to be waited for
+    :return: bool, if True the job has no tasks left
     """
-
     try:
 
-        cmd = f'qstat -u {username}'
-        process = subprocess.Popen(cmd, stdout=subprocess.PIPE, shell=True)
-        tmp = process.stdout.read().decode()
-        n_total = n_tasks(tmp, job_id)
+        time.sleep(10)
+        n_total = n_tasks(job_id)
         i = 31
         j = 6
+
         while n_total != 0:
             if i > 3:
 
-                running_process = subprocess.Popen(
-                    cmd + " -s r", stdout=subprocess.PIPE, shell=True)
-                running_tmp = running_process.stdout.read().decode()
-
-                if running_tmp != '':
-                    n_running = n_tasks(running_tmp, job_id)
-                else:
-                    n_running = 0
+                n_running = n_tasks(job_id, " -s r")
 
                 logger.info(f'{time.asctime(time.localtime())} - Job-ID {job_id}: '
                             f'{n_total} entries in queue. '
@@ -87,31 +83,51 @@ def wait_for_cluster(job_id):
                 j += 1
 
             if j > 7:
-                logger.info(str(tmp))
+                n_tasks(job_id, print_full=True)
                 j = 0
 
             time.sleep(30)
             i += 1
-            process = subprocess.Popen(cmd, stdout=subprocess.PIPE, shell=True)
-            tmp = process.stdout.read().decode()
-            n_total = n_tasks(tmp, job_id)
+            n_total = n_tasks(job_id)
 
-    except KeyboardInterrupt:
-        delete_cmd = f'qdel {job_id}'
-        process = subprocess.Popen(delete_cmd, stdout=subprocess.PIPE, shell=True)
-        msg = process.stdout.read().decode()
-        logger.info(str(msg))
+    finally:
+        # check again if there are still jobs left which might be true in case of raised exception
+        n_left = n_tasks(job_id)
+
+        # if there are jobs left, delete them if desired by user
+        if n_left > 0:
+
+            inpt = input(f'Delete remaining tasks of job {job_id}?: ')
+            if inpt in ['y', 'yes']:
+                logger.info(f'deleting job {job_id}')
+                delete_cmd = f'qdel {job_id}'
+                os.system(delete_cmd)
+                return True
+
+            else:
+                return False
+
+        else:
+            return True
 
 
-def n_tasks(tmp, job_id):
+def n_tasks(job_id=None, flags='', print_full=False):
     """
-    Returns the number of tasks given the output of qsub
-    :param tmp: output of qsub
-    :param job_id: int, optional, if given only tasks belonging to this job will we counted
-    :return: int
+    Returns the number of tasks for a given job ID
+    :param job_id: int, if given only tasks belonging to this job will we counted
+    :param flags: str, optional, flags to be added to qstat
+    :param print_full: bool, if True the full output of qstat will be printed
+    :return: int, number of tasks left
     """
+
+    process = subprocess.Popen(cmd+flags, stdout=subprocess.PIPE, shell=True)
+    tmp = process.stdout.read().decode()
+
     st = str(tmp)
     ids = np.array([int(s.split(' ')[2]) for s in st.split('\n')[2:-1]])
+
+    if print_full:
+        logger.info(st)
 
     if job_id:
         return len(ids[ids == job_id])
